requirements.txt
streamlit>=1.20.0
pandas>=1.5.0
plotly>=5.10.0
Pillow>=9.0.0        # pour charger et encoder votre logo.webp
openpyxl>=3.0.0      # pour lire les fichiers .xlsx via pandas
